+++ crates/ruma-events/src/room/message.rs
@@ -27,6 +27,8 @@ mod audio;
 mod content_serde;
 mod emote;
 mod file;
+#[cfg(feature = "unstable-msc4274")]
+mod gallery;
 mod image;
 mod key_verification_request;
 mod location;
@@ -47,6 +49,8 @@ mod without_relation;
 pub use self::audio::{
     UnstableAmplitude, UnstableAudioDetailsContentBlock, UnstableVoiceContentBlock,
 };
+#[cfg(feature = "unstable-msc4274")]
+pub use self::gallery::{GalleryItemType, GalleryMessageEventContent};
 #[cfg(feature = "unstable-msc4095")]
 pub use self::url_preview::{PreviewImage, PreviewImageSource, UrlPreview};
 pub use self::{
@@ -445,6 +449,10 @@ pub enum MessageType {
     /// A file message.
     File(FileMessageEventContent),
 
+    /// A media gallery message.
+    #[cfg(feature = "unstable-msc4274")]
+    Gallery(GalleryMessageEventContent),
+
     /// An image message.
     Image(ImageMessageEventContent),
 
@@ -499,6 +507,8 @@ impl MessageType {
             "m.audio" => Self::Audio(deserialize_variant(body, data)?),
             "m.emote" => Self::Emote(deserialize_variant(body, data)?),
             "m.file" => Self::File(deserialize_variant(body, data)?),
+            #[cfg(feature = "unstable-msc4274")]
+            "dm.filament.gallery" => Self::Gallery(deserialize_variant(body, data)?),
             "m.image" => Self::Image(deserialize_variant(body, data)?),
             "m.location" => Self::Location(deserialize_variant(body, data)?),
             "m.notice" => Self::Notice(deserialize_variant(body, data)?),
@@ -566,6 +576,8 @@ impl MessageType {
             Self::Audio(_) => "m.audio",
             Self::Emote(_) => "m.emote",
             Self::File(_) => "m.file",
+            #[cfg(feature = "unstable-msc4274")]
+            Self::Gallery(_) => "dm.filament.gallery",
             Self::Image(_) => "m.image",
             Self::Location(_) => "m.location",
             Self::Notice(_) => "m.notice",
@@ -583,6 +595,8 @@ impl MessageType {
             MessageType::Audio(m) => &m.body,
             MessageType::Emote(m) => &m.body,
             MessageType::File(m) => &m.body,
+            #[cfg(feature = "unstable-msc4274")]
+            MessageType::Gallery(m) => &m.body,
             MessageType::Image(m) => &m.body,
             MessageType::Location(m) => &m.body,
             MessageType::Notice(m) => &m.body,
@@ -616,6 +630,8 @@ impl MessageType {
             Self::Audio(d) => Cow::Owned(serialize(d)),
             Self::Emote(d) => Cow::Owned(serialize(d)),
             Self::File(d) => Cow::Owned(serialize(d)),
+            #[cfg(feature = "unstable-msc4274")]
+            Self::Gallery(d) => Cow::Owned(serialize(d)),
             Self::Image(d) => Cow::Owned(serialize(d)),
             Self::Location(d) => Cow::Owned(serialize(d)),
             Self::Notice(d) => Cow::Owned(serialize(d)),
@@ -676,6 +692,8 @@ impl MessageType {
                 }
                 MessageType::Audio(m) => (&mut m.body, None),
                 MessageType::File(m) => (&mut m.body, None),
+                #[cfg(feature = "unstable-msc4274")]
+                MessageType::Gallery(m) => (&mut m.body, None),
                 MessageType::Image(m) => (&mut m.body, None),
                 MessageType::Location(m) => (&mut m.body, None),
                 MessageType::ServerNotice(m) => (&mut m.body, None),
@@ -718,6 +736,8 @@ impl MessageType {
                 }
                 MessageType::Audio(m) => (&mut m.body, None),
                 MessageType::File(m) => (&mut m.body, None),
+                #[cfg(feature = "unstable-msc4274")]
+                MessageType::Gallery(m) => (&mut m.body, None),
                 MessageType::Image(m) => (&mut m.body, None),
                 MessageType::Location(m) => (&mut m.body, None),
                 MessageType::ServerNotice(m) => (&mut m.body, None),